requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.6
django-ipware==6.0.0
django-rest-authemail==2.1.7
djangorestframework==3.14.0
idna==3.6
phonenumberslite==8.13.26
psycopg2-binary==2.9.9
python-ipware==2.0.0
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.1.0
